Backend/requirements.txt
Flask
Flask-Cors
Pillow==10.0.1
face-recognition==1.3.0
numpy==1.24.3
opencv-python-headless
gunicorn
fastapi
uvicorn
python-multipart
pydantic==2.12.5
pydantic-settings==2.3.4
sqlalchemy==2.0.44
Flask-SQLAlchemy
Flask-Bcrypt
Flask-JWT-Extended